Fully implement AWS4 signing support
This fully implemeents AWS4 signing support, which was only partially added by 5d9165.

This adds the Sign field to Region, which can be used to sign a request with default sign method for that region.

In addition we have a single method that implements signing of requests for each supported signing method:
* SignV2
* SignV4
* SignRoute53

Fix V4Signer adding headers to the request after calculating the canonicalHeaders, due to this change all the tests have had to be updated with corrected checksums and headers.

Unfortunately it was not possible to make this change 100% backwards compatible.

The following changes are API incompatible and hence may require users to update their code:
* Removes Signature constants, as they are not used any more
* Removes ServiceInfo type, as its not used any more
* Removes Signer interface, as its not compatible with all signing methods.
* Changes the type of Region.RDSEndpoint from ServiceInfo -> string
* Renames Region.CloudWatchServicepoint -> CloudWatchEndpoint and changes its type from ServiceInfo -> string.
* Changes NewService from (auth Auth, serviceInfo ServiceInfo) -> (auth Auth, endpoint string, region Region).
* Changes NewV2Signer from (auth Auth, service ServiceInfo) -> (auth Auth, endpoint string)
* Changes NewCloudWatch from (auth aws.Auth, region aws.ServiceInfo) -> NewCloudWatch(auth aws.Auth, region aws.Region)
* Changes CanonicalRequest return from string to string, error

Also switch to using http status code constants while I'm here to ease comprehension.